model/sarimax/requirements.txt
aiohttp==3.7.4.post0
alpha-vantage==2.3.1
appdirs==1.4.3
async-timeout==3.0.1
attrs==20.3.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Cors==3.0.10
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
multidict==5.1.0
numpy==1.20.1
packaging==20.3
pandas==1.2.3
patsy==0.5.1
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.2
six==1.14.0
statsmodels==0.12.2
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
yarl==1.6.3
